backend-python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cassandra-driver==3.28.0
python-dotenv==1.0.0
gunicorn==21.2.0